Move Support\ProviderRepository to Core namespace
